requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
scikit-learn==1.5.1
numpy==2.0.1
python-dotenv==1.0.1
